+++ core-ng/src/main/java/core/framework/api/web/exception/BadRequestException.java
@@ -6,18 +6,30 @@ import core.framework.api.log.Warning;
  * @author neo
  */
 @Warning
-public class BadRequestException extends RuntimeException {
+public class BadRequestException extends RuntimeException implements ErrorCode {
+    public static final String DEFAULT_ERROR_CODE = "BAD_REQUEST";
+
     private static final long serialVersionUID = -2304226404736886782L;
+
+    private final String errorCode;
 
     public BadRequestException(String message) {
         super(message);
+        errorCode = DEFAULT_ERROR_CODE;
     }
 
-    public BadRequestException(String message, Throwable cause) {
+    public BadRequestException(String message, String errorCode) {
+        super(message);
+        this.errorCode = errorCode;
+    }
+
+    public BadRequestException(String message, String errorCode, Throwable cause) {
         super(message, cause);
+        this.errorCode = errorCode;
     }
 
-    public BadRequestException(Throwable cause) {
-        super(cause);
+    @Override
+    public String errorCode() {
+        return errorCode;
     }
 }