+++ core-ng/src/main/java/core/framework/api/web/exception/ErrorCode.java
@@ -1,0 +1,8 @@
+package core.framework.api.web.exception;
+
+/**
+ * @author neo
+ */
+public interface ErrorCode {
+    String errorCode();
+}