+++ core-ng/src/main/java/core/framework/api/web/exception/MethodNotAllowedException.java
@@ -7,10 +7,15 @@ import core.framework.api.log.Warning;
  * @author neo
  */
 @Warning
-public class MethodNotAllowedException extends RuntimeException {
+public class MethodNotAllowedException extends RuntimeException implements ErrorCode {
     private static final long serialVersionUID = 2349080664326196294L;
 
     public MethodNotAllowedException(HTTPMethod method) {
         super("method not allowed, method=" + method);
+    }
+
+    @Override
+    public String errorCode() {
+        return "METHOD_NOT_ALLOWED";
     }
 }