+++ core-ng/src/main/java/core/framework/api/web/exception/NotFoundException.java
@@ -6,14 +6,29 @@ import core.framework.api.log.Warning;
  * @author neo
  */
 @Warning
-public class NotFoundException extends RuntimeException {
+public class NotFoundException extends RuntimeException implements ErrorCode {
+    private static final String DEFAULT_ERROR_CODE = "NOT_FOUND";
     private static final long serialVersionUID = 8663360723004690205L;
+
+    private final String errorCode;
 
     public NotFoundException(String message) {
         super(message);
+        errorCode = DEFAULT_ERROR_CODE;
     }
 
-    public NotFoundException(Throwable cause) {
-        super(cause);
+    public NotFoundException(String message, String errorCode) {
+        super(message);
+        this.errorCode = errorCode;
+    }
+
+    public NotFoundException(String message, String errorCode, Throwable cause) {
+        super(message, cause);
+        this.errorCode = errorCode;
+    }
+
+    @Override
+    public String errorCode() {
+        return errorCode;
     }
 }