+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * properties: read in utf-8
 * httpclient: fix text() to use utf-8 as encoding
 * webservice: unified web service call error flow, RemoteServiceException provided structural info to be handled in client service app
+* exception: provide ErrorCode support for error flow.
 TODO: refactory i18n template
 
 ### 3.7.3 (10/27/2015)